UI changes, smaller take picture btn, add fab button to take new picture
Some UI improvements. Smaller "take picture" button in CameraActivity. Added floating action button in result activity which takes back to the camera activity to take a new picture. Added NavigatorActivity which is the starting activity, if a picture is previously taken is starts result activity, if it isn't it start camera activity