First Prototype of AllergensActivity
the activity has an editText and a ListView that shows allergen names and a button (that allows the user to select the allergen) in each row, the ListView is constructed thanks to the AllergenListAdapter Class. Access to the activity is done through an item added to the top-right menu of ResultActivity.
You can check the buttons' functionality through the Logs