Added swipeView feature to AllergensActivity
To add this new function the main AllergensActivity now only initializes the two Fragments that are used to make the layout. They are the AllergensSearchFragment and UserSAllergensFragment, and they are used to let the user add new allergens and view his own list separately to improve usability. You can either select the tab from the upper bar or swipe to switch between the two fragments.