Redundance improvement
I added the class BaseActivity. This class defines the behavior for activities that have a bottomnavigationView. These activities should extend this one and then define a BottomNavigationView with "bottom_navigation" as its id in their layout xml file; all the instructions that manage the bottom menu are defined in BaseActivity and made the same for all activities, removing the need to define a BottomNavigationView and the cases for each button every time, reducing as such the redundancy of the code.