Final code commit
Here i put 6 classes to review. 
Brightness recognition is used in ResultActivity to determine under/over exposure.
AllergensListAdapter, AllergensSearchFragment, UserAllergensFragment and MainAllergensActivity let the user select a personal list of allergens that are looked for in the list of recognized ingredients in ResultActivity.
BaseActivity is used to improve (decrease) code redundancy for the implementation of a bottomBarMenu that lets the user switch between activities: in the previous iterations (that i will not post here, there also was a version of an hamburgermenu that did the same thing as the bottombarmenu), every activity had to re-define a bottombar and all the cases for each button in the menu. Now an activity that should have a bottombarmenu only has to extend BaseActivity and implement two of its methods.